requirements.txt
openai
langchain
tiktoken
faiss-cpu
python-dotenv
unstructured[all-docs]
fastapi
uvicorn